refactor/#6 [style] : 코드 리팩토링 과정 중에, 변경된 파일 싹 정리
