+++ app/src/main/java/com/devnuts/ruflu/MainAdapter.kt
@@ -1,5 +1,6 @@
 package com.devnuts.ruflu
 
+import android.util.Log
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
