--- app/src/main/java/com/devnuts/ruflu/RufluApplication.kt
@@ -1,8 +1,0 @@
-package com.devnuts.ruflu
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-//@HiltAndroidApp
-//class RufluApplication : Application() {
-//}
