+++ app/src/main/java/com/devnuts/ruflu/data/source/MatchDataSource.kt
@@ -1,6 +1,6 @@
 package com.devnuts.ruflu.data.source
 
-import com.devnuts.ruflu.data.api.response.card.UserModel
+import com.devnuts.ruflu.data.api.response.home.model.UserModel
 import retrofit2.Call
 
 interface MatchDataSource {
