+++ app/src/main/java/com/devnuts/ruflu/domain/repository/ChatRepository.kt
@@ -1,14 +1,14 @@
 package com.devnuts.ruflu.domain.repository
 
 import com.devnuts.ruflu.ui.model.chat.ChatRoom
-import com.devnuts.ruflu.util.RufluApp
-import com.devnuts.ruflu.util.ServerAPI
+import com.devnuts.ruflu.RufluApp
+import com.devnuts.ruflu.util.RufluApiService
 import retrofit2.Call
 
 class ChatRepository {
-    private val api = RufluApp.retrofit.create(ServerAPI::class.java)
+    //private val api = RufluApp.retrofit.create(RufluApiService::class.java)
 
-    fun getMyChattingRoomList(): Call<List<ChatRoom>> {
-        return api.getMyChattingRoomList()
-    }
+//    fun getMyChattingRoomList(): Call<List<ChatRoom>> {
+//        //return api.getMyChattingRoomList()
+//    }
 }
