+++ app/src/main/java/com/devnuts/ruflu/domain/repository/MainRepository.kt
@@ -3,7 +3,6 @@ package com.devnuts.ruflu.domain.repository
 import retrofit2.Call
 
 interface MainRepository {
-
     fun updateUserLocation(latitude: Double, longitude: Double): Call<String>
 
     fun executeFcmServiceToken(token: String): Call<String>
