+++ app/src/main/java/com/devnuts/ruflu/ui/chat/fragment/ChatRoomFragment.kt
@@ -14,7 +14,7 @@ import com.devnuts.ruflu.databinding.ItemMyChatRoomBinding
 import com.devnuts.ruflu.ui.adapter.MassagesAdapter
 import com.devnuts.ruflu.ui.chat.viewmodel.ChatSharedViewModel
 import com.devnuts.ruflu.ui.model.chat.ChatMessage
-import com.devnuts.ruflu.util.RufluApp
+import com.devnuts.ruflu.RufluApp
 import com.github.nkzawa.emitter.Emitter
 import com.github.nkzawa.socketio.client.IO
 import com.github.nkzawa.socketio.client.Socket
@@ -44,7 +44,7 @@ class ChatRoomFragment : Fragment() {
 
     private fun init() {
         try {
-            mSocket = IO.socket(RufluApp.url + ":" + RufluApp.port)
+            //mSocket = IO.socket(RufluApp.url + ":" + RufluApp.port)
         } catch (e: URISyntaxException) {
             e.printStackTrace()
         }
