+++ app/src/main/java/com/devnuts/ruflu/ui/like/viewmodel/MatchViewModel.kt
@@ -2,9 +2,8 @@ package com.devnuts.ruflu.ui.like.viewmodel
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.devnuts.ruflu.data.api.response.card.UserModel
+import com.devnuts.ruflu.data.api.response.home.model.UserModel
 import com.devnuts.ruflu.domain.repository.MatchRepository
-import com.devnuts.ruflu.domain.repository.SomeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import retrofit2.Call
 import retrofit2.Callback
