+++ app/src/main/java/com/devnuts/ruflu/worker/AddCookiesInterceptor.kt
@@ -1,6 +1,5 @@
 package com.devnuts.ruflu.worker
 
-import com.devnuts.ruflu.util.RufluApp
 import okhttp3.Interceptor
 import okhttp3.Response
